refactor: removing redeclaration constants
