+++ constants/constants.ts
@@ -1,0 +1,5 @@
+const DOCUMENT_WRITE_SIZE = 500;
+
+module.exports = {
+    DOCUMENT_WRITE_SIZE,
+};
